% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtwittertext.R
\name{add_twitter_text}
\alias{add_twitter_text}
\title{Create HTML/Markdown text with Twitter's logo and handle/text}
\usage{
add_twitter_text(twitter_handle, logo_size = 9, logo_url = NULL)
}
\arguments{
\item{twitter_handle}{like @abcd}

\item{logo_size}{(default = 9)}

\item{logo_url}{custom image (i.e dark mode). If none provide, will download logo from wikimedia}
}
\value{
text
}
\description{
Create HTML/Markdown text with Twitter's logo and handle/text
}
