% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplottheme.R
\name{custom_map_theme}
\alias{custom_map_theme}
\title{Custom theme for ggplot maps}
\usage{
custom_map_theme(google_font = "Roboto", map_title_size = 14)
}
\arguments{
\item{google_font}{font, from Google Fonts}

\item{map_title_size}{Title's size}
}
\value{
ggplot theme object
}
\description{
Custom theme for ggplot maps
}
