% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custommapthememd.R
\name{custom_map_theme_md}
\alias{custom_map_theme_md}
\title{Custom theme for ggplot maps, using ggtext for markdown text}
\usage{
custom_map_theme_md(
  google_font = "Roboto",
  map_title_size = 14,
  map_subtitle_size = 12,
  map_caption_size = 9,
  legend_pos = "none",
  legend_dir = "horizontal",
  title_pos = "plot",
  background_colour = "white",
  plot_margin = c(3, 0, 1, 0)
)
}
\arguments{
\item{google_font}{font, from Google Fonts}

\item{map_title_size}{Title's size}

\item{map_subtitle_size}{Subtitle's size}

\item{map_caption_size}{Caption's size}

\item{legend_pos}{legend's position (default none)}

\item{legend_dir}{legend's direction (defaul horizontal)}

\item{title_pos}{title's position (default plot)}

\item{background_colour}{background's overall colour (default white)}

\item{plot_margin}{margins c(top,right,bottom,left) - default c(3,0,1,0)}
}
\description{
Custom theme for ggplot maps, using ggtext for markdown text
}
