% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplottheme.R
\name{custom_plot_theme}
\alias{custom_plot_theme}
\title{Custom theme for ggplot plots}
\usage{
custom_plot_theme(
  google_font = "Roboto",
  title_colour = "#272928",
  general_colour = "azure4",
  background_colour = "#fcf8f7",
  legend_pos = "right",
  legend_dir = "vertical"
)
}
\arguments{
\item{google_font}{font, from Google Fonts}

\item{title_colour}{Colour for title and legend textx}

\item{general_colour}{Colour for all other texts in plot}

\item{background_colour}{Background colour}

\item{legend_pos}{legend's position (default right)}

\item{legend_dir}{legend's direction (defaul vertical)}
}
\value{
ggplot theme object
}
\description{
Custom theme for ggplot plots
}
