% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplotthememd.R
\name{custom_plot_theme_md}
\alias{custom_plot_theme_md}
\title{Custom theme for ggplot plots with markdown texts}
\usage{
custom_plot_theme_md(
  google_font = "Roboto",
  title_colour = "#272928",
  general_colour = "azure4",
  background_colour = "white",
  legend_pos = "right",
  legend_dir = "vertical",
  plot_margin = c(3, 0, 1, 0),
  title_size = 14,
  subtitle_size = 12,
  caption_size = 9,
  axis_size = 10,
  legend_size = 8
)
}
\arguments{
\item{google_font}{font, from Google Fonts}

\item{title_colour}{Colour for title and legend textx}

\item{general_colour}{Colour for all other texts in plot}

\item{background_colour}{Background colour}

\item{legend_pos}{legend's position (default right)}

\item{legend_dir}{legend's direction (defaul vertical)}

\item{plot_margin}{margins c(top,right,bottom,left) - default c(3,0,1,0)}

\item{title_size}{Title's size}

\item{subtitle_size}{Subtitle's size}

\item{caption_size}{Caption's size}

\item{axis_size}{text size for axis items}

\item{legend_size}{text size for legend}
}
\value{
ggplot theme object
}
\description{
Custom theme for ggplot plots with markdown texts
}
