% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customstriptheme.R
\name{custom_strip_theme}
\alias{custom_strip_theme}
\title{Custom theme for temperatue strips
based on https://dominicroye.github.io/en/2018/how-to-create-warming-stripes-in-r/, modified}
\usage{
custom_strip_theme(google_font = "Roboto", rbrewer_pal = "RdBu")
}
\arguments{
\item{google_font}{font, from Google Fonts}

\item{rbrewer_pal}{RColorBrewer palette used}
}
\value{
ggplot theme object
}
\description{
Custom theme for temperatue strips
based on https://dominicroye.github.io/en/2018/how-to-create-warming-stripes-in-r/, modified
}
