% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitsformatter.R
\name{digits_formatter}
\alias{digits_formatter}
\title{Small wrapper of formatC for number formatting}
\usage{
digits_formatter(x, sig_digits = 3, nbr_format = "f")
}
\arguments{
\item{x}{number to format}

\item{sig_digits}{decimal precission}
}
\value{
formatted number
}
\description{
Small wrapper of formatC for number formatting
}
