% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveimage.R
\name{save_image}
\alias{save_image}
\title{save ggplot into PNG file}
\usage{
save_image(
  plot_name,
  file_name,
  dpi_value = 600,
  width_value = NA,
  height_value = NA,
  unit_value = NA
)
}
\arguments{
\item{plot_name}{ggplot object}

\item{file_name}{file's name (including .png)}

\item{dpi_value}{resolution (default 600)}

\item{width_value}{width for ggsave}

\item{height_value}{height for ggsave}

\item{unit_value}{unit for ggsave}
}
\value{
saved file in disk

ggplot theme object
}
\description{
save ggplot into PNG file
}
