% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplottheme.R
\name{set_plot_colours}
\alias{set_plot_colours}
\title{Set colour for standard chart fillings and lines
@return no output
@import ggplot2
@export set_plot_colours}
\usage{
set_plot_colours()
}
\description{
Set colour for standard chart fillings and lines
@return no output
@import ggplot2
@export set_plot_colours
}
