% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_levy_facet.R
\name{theme_levy_facet}
\alias{theme_levy_facet}
\title{theme_levy_facet}
\usage{
theme_levy_facet(base_size = 12, font = "Source Sans Pro")
}
\arguments{
\item{base_size}{12 by default}

\item{font}{"Source Sans Pro" by default. \code{extrafont::fonttable()[["FamilyName"]]} to see what's available}
}
\value{
ggplot theme
}
\description{
theme_levy_facet
}
\details{

}
\examples{
\dontrun{
ggplot(data.frame(x = 1:10, y = rnorm(10), z = sample(letters[1:2])),
  aes(x, y)) +
  geom_point() +
  facet_wrap(~ z) +
  theme_levy_facet()
}
}

