% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_levy_single.R
\name{theme_levy_single}
\alias{theme_levy_single}
\title{theme_levy_single}
\usage{
theme_levy_single(base_size = 12, font = "Source Sans Pro")
}
\arguments{
\item{base_size}{12 by default}

\item{font}{"Source Sans Pro" by default. \code{extrafont::fonttable()[["FamilyName"]]} to see what's available}
}
\value{
ggplot theme
}
\description{
theme_levy_single
}

