% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveimage.R
\name{save_image}
\alias{save_image}
\title{save ggplot into PNG file}
\usage{
save_image(plot_name, file_name, dpi_value = 600)
}
\arguments{
\item{plot_name}{ggplot object}

\item{file_name}{file's name (including .png)}

\item{dpi_value}{resolution (default 600)}
}
\value{
saved file in disk

ggplot theme object
}
\description{
save ggplot into PNG file
}
