% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_levy.R
\name{theme_levy}
\alias{theme_levy}
\title{theme_levy}
\usage{
theme_levy(base_size = 12, font = "Source Sans Pro")
}
\arguments{
\item{base_size}{12 by default}

\item{font}{"Source Sans Pro" by default. \code{extrafont::fonttable()[["FamilyName"]]} to see what's available}
}
\value{
ggplot theme
}
\description{
theme_levy
}
\examples{
\dontrun{
ggplot(data.frame(x = 1:10, y = rnorm(10)), aes(x, y)) +
  geom_point() +
  theme_levy()
  }
}

